fix: Ollama crash recovery in call_api.py, simplify run_experiments.py
- call_api.py: add _ensure_ollama() singleton with retry logic (3 attempts + restart)
- call_api.py: Model.__init__ no longer kills existing Ollama server
- run_experiments.py: remove duplicate _ensure_ollama(), replace with _check_ollama()
- All 46 tests passing
