+++ scripts/run_experiments.py
@@ -5,7 +5,6 @@ Usage:
     python run_experiments.py --benchmarks cv-screening,stereoset --max-samples 10 --timeout 600
 """
 import argparse
-import atexit
 import json
 import logging
 import os
@@ -35,29 +34,16 @@ SLM_MODELS = [
 ]
 
 
-_ollama_server = None
-
-
-def _ensure_ollama():
-    """Check if Ollama is responding; restart if not (singleton)."""
-    global _ollama_server
+def _check_ollama():
+    """Quick check that Ollama is responding."""
     try:
         subprocess.run(
             ["ollama", "list"], capture_output=True, check=True, timeout=10
         )
+        return True
     except Exception:
-        # Stop old server and unregister its atexit handler
-        if _ollama_server is not None:
-            atexit.unregister(_ollama_server.stop)
-            try:
-                _ollama_server.stop()
-            except Exception:
-                pass
-        logger.warning("Ollama not responding, restarting...")
-        from slm_bias_testing.ollama_setup import OllamaServer
-        _ollama_server = OllamaServer(kill_existing=True)
-        _ollama_server.start()
-        logger.info("Ollama restarted")
+        logger.error("Ollama not responding — run 'ollama serve' first")
+        return False
 
 
 def run_benchmarks(models, benchmarks, output_dir, max_samples, timeout):
@@ -85,8 +71,10 @@ def run_benchmarks(models, benchmarks, output_dir, max_samples, timeout):
                 results_summary.append(summary)
                 continue
 
-            # Ensure Ollama is alive before running the benchmark
-            _ensure_ollama()
+            # Check Ollama is alive before running the benchmark
+            if not _check_ollama():
+                logger.error("Skipping %s / %s — Ollama not available", model_name, benchmark)
+                continue
 
             cmd = [
                 sys.executable, "-m", "slm_bias_testing.runner",
